mgie_api/requirements_mgie_api.txt
fastapi[standard]
uvicorn
pydantic
pillow
